requirements.txt
requests~=2.31.0
influxdb~=5.3.1
schedule~=1.2.1
pycryptodomex~=3.20.0